github: do not fail on push
In case this script is invoked outside a pull request, the event
structure is different, and there will not be a pull_request member. As
this script might be used in automation for e.g. pushes, this use case
should be supported as well. Hence, this change implements checking for
the presence of the member first, and handling the cases accordingly.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>